4.21 added

+++ part4/controllers/Blogs.js
@@ -12,19 +12,16 @@ blogRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-
-
 blogRouter.post('/', async (request, response) => {
     const body = request.body
-    const token = req.token
-    console.log("token: ", token)
+    const token = request.token
     const decodedToken = jwt.verify(token, process.env.SECRET)
     if (!token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
     //const user = await User.findById(body.userID)
-    console.log("hweere ", user)
+
 
     const blog = new Blog({
         title: body.title,
@@ -46,10 +43,18 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (req, res, next) => {
+    const token = req.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
     try {
-        const result = await Blog.findByIdAndRemove(req.params.id)
-        res.status(204).end()
+        const blog = await Blog.findById(req.params.id)
+        if (blog.user.toString() === decodedToken.id.toString()) {
+            await Blog.findByIdAndRemove(req.params.id)
+        }
+        res.status(204).json("deleted succesfully")
     }
     catch (error) {
         next(error)
